+++ src/Options.hs
@@ -1,6 +1,0 @@
-module Options where
-
-newtype Options = Options
-  { optVerbose :: Bool
-  }
-  deriving (Show)