+++ src/main.hs
@@ -1,37 +1,36 @@
-{-# LANGUAGE ApplicativeDo #-}
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE RecordWildCards #-}
+module Main where
 
-import Data.Foldable (fold)
+import Common.Options (subparsers)
+import Control.Monad (join)
 import Data.Version (showVersion)
-import Options (Options (..))
 import Options.Applicative
 import PackageInfo_cardano_ledger_release_tool (version)
 
-import qualified Changelogs
+import qualified Cabal.Command as Cabal
+import qualified Changelogs.Command as Changelogs
 import qualified System.Console.Terminal.Size as TS
+import qualified Workflow.Command as Workflow
 
 main :: IO ()
 main = do
   cols <- maybe 100 TS.width <$> TS.size
 
-  (subcmd, opts) <-
-    customExecParser
-      (prefs $ columns cols)
-      ( info
-          ( helper <*> simpleVersioner (showVersion version) <*> do
-              optVerbose <-
-                switch $
-                  help "Produce verbose output"
-                    <> short 'v'
-                    <> long "verbose"
-              subcmd <-
-                subparser . fold $
-                  [ Changelogs.subcmd
-                  ]
-              pure (subcmd, Options {..})
-          )
-          (fullDesc <> header "Cardano Ledger release tool")
-      )
+  let
+    versionOption v =
+      infoOption (showVersion v) $
+        help "Show version information"
+          <> short 'V'
+          <> long "version"
+          <> hidden
 
-  subcmd opts
+  join . customExecParser (prefs $ columns cols) $
+    info
+      ( helper
+          <*> versionOption version
+          <*> subparsers
+            [ Cabal.subcmd
+            , Changelogs.subcmd
+            , Workflow.subcmd
+            ]
+      )
+      (fullDesc <> header "Cardano Ledger release tool")