+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 Version history for `cardano-ledger-release-tool`
 
+## 0.2.0.0
+
+* Add a `cabal` group of subcommands
+* Add a `workflow` group of subcommands
+* Make the `changelogs` subcommand a group
+* Rename the executable to `cleret`
+* Add `-V` as a short-option form of `--version`
+
 ## 0.1.1.0
 
 * Add a `--version` option